# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAphoto
  description: |-
    OpenAPI document for the WASAPhoto application
  version: 1.1.0
...
paths:
...
  /profiles/{userID}/followings/{targetUserID}:
    parameters:
    - $ref: "#/components/parameters/userID"
    - $ref: "#/components/parameters/targetUserID"
    put:
      summary: Follow a user
      description: Add the user that you want to follow into list of following
      tags: [ "User" ]
      operationId: followUser
      responses:
        "200":
          description: The user followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: Unfollow a user
      tags: [ "User" ]
      operationId: unfollowUser
      description: Unfollow the user with the given userID and delete it from the following
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /profiles/{userID}/followers:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      summary: Get followers
      tags: [ "User" ]
      operationId: listFollowers
      description: Get the list of users that are following the user.
      responses:
        "200":
          description: List of followers retrieved
          content:
            application/json:
              schema:
                description: The list of followers
                type: object
                properties:
                  followers:
                    description: The list of users who follow the specified user
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /profiles/{userID}/bans/{targetUserID}:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/targetUserID"
    put:
      summary: Ban a user
      tags: [ "User" ]
      operationId: banUser
      description: Ban the user with the given userID and delete it from the following
      responses:
        "200":
          description: User banned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Unban a user
      tags: [ "User" ]
      operationId: unbanUser
      description: Unban the user with the given userID and delete it from the following
      responses:
        "200":
          description: User unbanned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /profiles/{userID}/bans:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      summary: Get banned users
      tags: [ "User" ]
      operationId: getBannedList
      description: Get the list of users that are banned by the user.
      responses:
        "200":
          description: List of banned users retrieved
          content:
            application/json:
              schema:
                description: The list of banned users
                type: object
                properties:
                  banned:
                    description: The list of banned users
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /profiles/{userID}/posts/{photoID}:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/photoID"

    delete:
      summary: Delete a post
      tags: [ "Posts" ]
      operationId: deletePhoto
      description: Delete the post with the given photoID.
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /profiles/{userID}/posts/{photoID}/comments:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/photoID"
    post:
      summary: Comment on a post
      tags: [ "Posts" ]
      operationId: commentPhoto
      description: Comment on the post with the given photoID.
      requestBody:
        description: The caption of the comment
        content:
          application/json:
            schema:
              description: The caption of the comment posted
              type: object
              properties:
                caption:
                  $ref: "#/components/schemas/Comment/properties/lyric"
      responses:
        "201":
          description: Comment added successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      summary: Get comments on a post
      tags: [ "Posts" ]
      operationId: getComments
      description: Get the list of comments on the post with the given photoID.
      responses:
        "200":
          description: List of comments retrieved
          content:
            application/json:
              schema:
                description: The list of comments
                type: object
                properties:
                  comments:
                    description: The list of comments posted by the user
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    Username:
      title: Username
      description: The username of the user
      type: object
      properties:
        username:
          type: string
          pattern: "^[a-z0-9]{3,13}$"
          description: User's username
          example: "john123"
          minLength: 3
          maxLength: 13
    ImageData:
      title: Image
      description: The binary of the image (1-20Mb)
      type: string
      format: binary
      minLength: 1
      maxLength: 20971520
    Caption:
      title: Caption
      description: The caption of the photo
      type: string
      pattern: '^[^\/\\]{0,64}$'
      example: "This is a caption"
      minLength: 0
      maxLength: 64
    Timestamp:
      title: Timestamp
      description: The timestamp of the post.
      type: string
      format: date-time
      pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}$"
      example: "2020-12-31T23:59:59"
      minLength: 19
      maxLength: 19
    UserID:
      type: integer
      description: The user id of the user.
      example: 100
    User:
      title: User
      description: The user object
      type: object
      properties:
        userID:
          $ref: "#/components/schemas/UserID"
        username:
          $ref: "#/components/schemas/Username"
        profilePicture:
          $ref: "#/components/schemas/ImageData"
...
    PhotoID:
      description: The ID of the post
      type: integer
      readOnly: true
      example: 100
...
    Comment:
      title: Comment
      description: The comment object of the post
      type: object
      properties:
        commentID:
          type: integer
          description: The comment id of the comment.
          readOnly: true
          example: 100
        ownerID:
          $ref: "#/components/schemas/UserID"
        photoID:
          $ref: "#/components/schemas/PhotoID"
        lyric:
          $ref: "#/components/schemas/Caption"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    Message:
      title: Message
      description: This object represent a generic description of a message
      type: object
      readOnly: true
      properties:
        error:
          description: The message of the error
          type: string
          pattern: '\S.+'
          example: Username is already taken
          minLength: 0
          maxLength: 200
  parameters:
    userID:
      name: userID
      in: path
      required: true
      description: The user id of the user.
      schema:
        $ref: "#/components/schemas/UserID"
    photoID:
      name: photoID
      in: path
      description: The numeric photo id
      required: true
      schema:
        $ref: "#/components/schemas/PhotoID"
...
    targetUserID:
      name: targetUserID
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UserID"
  responses:
...
    Success:
      description: Success
      content:
        application/json:
          schema:
            description: Success message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 201 response
...
    NoContent:
      description: No content
      content:
        application/json:
          schema:
            description: No content message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 400 response
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            description: Bad request message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 401 response
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            description: Unauthorized message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 403 response
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            description: Forbidden message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 404 response
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            description: Internal server error message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

# Apply the security global, to all the operations
security:
- bearerAuth: []